UPDATE the matrix now uses floats to represent heat. It still prints integers for readability, though.

+++ Matrix.c
@@ -3,15 +3,16 @@
 //
 
 #include <malloc.h>
+#include <stdio.h>
 #include "Matrix.h"
 
 /**
  * Init matrix representation.
  */
-int init_matrix(int size, int max_temp_value, matrix_representation * m){
+int init_matrix(int size, float max_temp_value, matrix_representation * m){
     m->size = size;
     m->middle_index = size/2;
-    m->matrix = malloc(sizeof(int*)*size);
+    m->matrix = malloc(sizeof(float*)*size);
     m->max_temp_value = max_temp_value;
     m->matrix[m->middle_index] = max_temp_value;
     return 0;
@@ -24,7 +25,7 @@ void print_matrix(matrix_representation * m){
     printf("Matrix : \n");
     int i;
     for(i = 0; i < m->size; i++){
-        printf("%d ", m->matrix[i]);
+        printf("%f ", m->matrix[i]);
     }
     printf("\n-------------\n");
 }
