+++ Matrix2D.h
@@ -5,15 +5,15 @@
 #ifndef PROG_CONC_DIFF_MATRIX2D_H
 typedef struct {
     int size;
-    int ** matrix;
+    float ** matrix;
     int middle_index;
     int half_range;
-    int max_temp_value;
+    float max_temp_value;
 } matrix_2d;
 
 
 
-int init_matrix_2d(int, int, matrix_2d *);
+int init_matrix_2d(int, float, matrix_2d *);
 
 void print_matrix_2d(matrix_2d *);
